Crate library

Independent and project agnostic libraries

Ideally, any of the library submodules in this module can be extracted into their own crate at any given time (like it happened with jatsl). Libraries in this module have been developed with WebGrid in mind and are powering core functionalities, however, they are in no way bound to the project and everything domain specific has been extracted into the domain module.

Modules

communication

Structures to communication between services in a distributed system

helpers

Various small helper functions

http

Structures for handling and sending HTTP requests

storage

Structures to persist and retrieve arbitrary binary data

Macros

make_responder_chain_service_fn

Combines a number of responder instances and creates a new hyper service function through the use of make_service_fn. To allow for concurrent access to the responders in the chain, they are wrapped in Arc pointer types.

responder_chain

Chains together a number of Responder implementations

Structs

AccumulatedPerformanceMetrics

Metrics about a process accumulated over a certain timespan

PerformanceMonitor

Tool to collect AccumulatedPerformanceMetrics for a given process

Enums

PerformanceMonitoringTarget

Type of target to monitor the performance of

Type Definitions

BoxedError

Generic error type

EmptyResult

Result with no value and a BoxedError