Module library::communication::event

Structures to realise an event-driven service architecture

In an event driven world, services have no knowledge of each other. Each service operates independently and during the operation, certain events occur. For each of these (that are of relevance to other services) an event Notification is published. Every interested party may then subscribe to the Queue of notifications for a specific event and process them. While processing, it is common that other events are triggered and thus further notifications published.

Additionally, notifications are consumed in a reliable and resilient way using a concept called ConsumerGroups. Instead of using simple publish subscribe between all connected services, messages are stored in a log-like data structure (usually of limited length where old elements are evicted).

When reading from this data structure, services may define a QueueLocation from which they want to begin processing. All notifications have to be acknowledged once processing concludes. Upon crashing, the Consumer may then resume from the last acknowledged notification. This ensures that no QueueEntries are left unprocessed.

The last benefit of this data structure is that multiple Consumers may share a ConsumerGroup. All participants in a group then collectively process the incoming notification stream where each notification is assigned to only one consumer within the group (effectively implementing load balancing and simple, dynamic scalability).

Structs

ConsumerGroupDescriptor

Definition of a consumer group

NotificationFrame

Frame around a notification containing additional context and information

QueueDescriptor

Describes a notification queue and its parameters

Enums

ConsumerGroupIdentifier

Unique identifier for a group of consumers

QueueLocation

Location within the queue

Traits

Consumer

Entity which may consume and process Notifications

ConsumerExt

Helper functions to aid the consumption of messages

Notification

Entity to notify other service about an event that took place

NotificationPublisher

Publisher for Notifications

QueueEntry

Useful functions for QueueEntry implementations with default implementations

QueueProvider

Allows consumption of notification queues using consumer groups

RawNotificationPublisher

Structure which allows publishing of serialized data into a queue

RawQueueEntry

Entry retrieved from a Queue providing a raw payload

Type Definitions

ConsumerIdentifier

Unique identifier of a consumer within a ConsumerGroup

QueueDescriptorExtension

Adds additional information to the key of a queue, specialising it further