Trait library::communication::event::ConsumerExt

pub trait ConsumerExt {
    fn recv_queue_item<'life0, 'async_trait, Q>(
        &'life0 self, 
        item: Result<<Q as QueueProvider>::Entry, BoxedError>, 
        seq_id: u64
    ) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>>
    where
        Q: QueueProvider + Send + Sync,
        Q: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;

    fn consume_queue<'life0, 'life1, 'life2, 'life3, 'async_trait, Q>(
        &'life0 self, 
        provider: Q, 
        group: &'life1 ConsumerGroupDescriptor, 
        consumer: &'life2 str, 
        extension: &'life3 Option<QueueDescriptorExtension>
    ) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>>
    where
        Q: QueueProvider + Send + Sync,
        Q: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;
}

Helper functions to aid the consumption of messages

Required methods

fn recv_queue_item<'life0, 'async_trait, Q>(
&'life0 self,
item: Result<<Q as QueueProvider>::Entry, BoxedError>,
seq_id: u64
) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>> where
Q: QueueProvider + Send + Sync,
Q: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Handles individual queue entries by parsing their payload and handind them to the underlying consumer.

fn consume_queue<'life0, 'life1, 'life2, 'life3, 'async_trait, Q>(
&'life0 self,
provider: Q,
group: &'life1 ConsumerGroupDescriptor,
consumer: &'life2 str,
extension: &'life3 Option<QueueDescriptorExtension>
) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>> where
Q: QueueProvider + Send + Sync,
Q: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Consumes notifications from a queue using the given provider and acknowledges those that have been successfully processed.