Module library::communication::request

Structures to realise a request-response pattern

This module provides types for a request-response pattern where requests are reliable and can be answered by multiple services. In contrast, responses are not fully crash-tolerant and will get lost if the requesting service crashes while processing it.

However, this is by design as a crash in a service usually results in the request (and thus response location) being lost as well. Despite this, it is assumed that requests will usually be triggered while processing a Notification which is inherently crash-resistant. Thus, when recovering, a service will repeat the processing of the notification and in turn repeat the Request. For this reason, requests may not have side effects.

When talking about the request-response pattern, there are two parties involved:

Requesting side

Responding side

On the requesting side, a Requestor struct is used to send a Request and receive the linked Request::Response. In its essence, it just wraps a NotificationPublisher (used to dispatch the request) and a ResponseCollector to wait for responses.

On the responding side, a Responder struct is used to wait for incoming Requests, by the means of implementing the Consumer trait, process them using a RequestProcessor and publishing the returned response using a ResponsePublisher.

Structs

CompositeRequestor

Requestor implementation by combining a NotificationPublisher and ResponseCollector

Responder

Convenience wrapper to process requests and send responses

Enums

RequestError

Error type for sending requests

ResponderError

Error that may be thrown while responding to a request

ResponseCollectionTimeout

Timeout structure for response collection

Traits

OptionalRequestProcessor

Structure which processes requests and may either produce a response or ignore the request

RawResponseCollector

Structure to wait for and collect one or more raw responses

RawResponsePublisher

Structure which allows publishing of raw responses

Request

Query for information which can be replied to

RequestProcessor

Structure which processes requests and produces responses

Requestor

Handler for sending requests and collecting responses

ResponseCollector

Collector of typed responses to Requests

ResponsePublisher

Publisher for responses to Requests

Type Definitions

ResponseLocation

Describes a location where a response should be sent to