Library

Function main

pub(crate) fn main()

The main entry point for the application.

This function initializes the command structure and executes the asynchronous function defined within it. The function is marked with the #[tokio::main] attribute to enable asynchronous execution using the Tokio runtime.

Panics

This function does not panic.

Example

#[tokio::main]
async fn main() { (Struct::Binary::Command::Struct::Fn().Fn)().await }